Reduce unnecessary console output validation in test_commands (#67)
We don't really need the pull behavior hint to validate what's we're
trying to do in these checks, so we can disable it in newer git clients
to achieve the same behavior across the board.

Additionally, the addition of the static global .gitconfig file means
that we don't need to worry about insteadOf mappings messing with our
git URLs, so that workaround can be dropped.

Signed-off-by: Scott K Logan <logans@cottsay.net>