Ignore nondeterministic line ordering in test_commands (#66)
These tests inspect the console output from vcs2l itself, and the order
in which the messages appears doesn't appear to be deterministic and can
vary by platform, or in some cases, each invocation.

Modern Python versions will generally maintain the order in which values
are added to a dictionary, which may be a preferable behavior here.
Rather than sort the values when they're printed from vcs2l, we can just
ignore the ordering in our checks.

Signed-off-by: Scott K Logan <logans@cottsay.net>