Fix detection of annotated git tags
This commit extends vcs2l's tag detection scheme to support annotated
tags, building on the existing support for lightweight tags. The new
detection logic adds awareness of some git plumbing details regarding
how the tag's message is stored in relation to the SHA it references.

I updated the test to use an annotated tag for one of the tags in the
list to exercise this feature.

Signed-off-by: Scott K Logan <logans@cottsay.net>