Add python3-vcs2l-vcstool-shim debian package
This sub-package presents vcs2l as a wholesale replacement for vcstool
and should satisfy any existing dependencies on vcstool without
requiring changes to vcs2l itself. It operates under the fact that the
existing vcs2l API is similar enough to vcstool that it can simply be
installed and used under that name.

When the vcs2l API deviates from vcstool enough, this commit may be
reverted and the shim package dropped.

Signed-off-by: Scott K Logan <logans@cottsay.net>